Refactor index updates to prepare for #3022
Related to #1099
