+++ janusgraph-core/src/main/java/org/janusgraph/graphdb/database/index/IndexRecords.java
@@ -1,0 +1,35 @@
+// Copyright 2022 JanusGraph Authors
+//
+// Licensed under the Apache License, Version 2.0 (the "License");
+// you may not use this file except in compliance with the License.
+// You may obtain a copy of the License at
+//
+//      http://www.apache.org/licenses/LICENSE-2.0
+//
+// Unless required by applicable law or agreed to in writing, software
+// distributed under the License is distributed on an "AS IS" BASIS,
+// WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+// See the License for the specific language governing permissions and
+// limitations under the License.
+
+package org.janusgraph.graphdb.database.index;
+
+import org.janusgraph.graphdb.database.IndexRecordEntry;
+import org.janusgraph.graphdb.database.util.IndexRecordUtil;
+
+import java.util.ArrayList;
+import java.util.Arrays;
+import java.util.stream.Collectors;
+
+public class IndexRecords extends ArrayList<IndexRecordEntry[]> {
+
+    @Override
+    public boolean add(IndexRecordEntry[] record) {
+        return super.add(Arrays.copyOf(record,record.length));
+    }
+
+    public Iterable<Object[]> getRecordValues() {
+        return this.stream().map(IndexRecordUtil::getValues).collect(Collectors.toList());
+    }
+
+}
